Get download working in a basic sense (#1)

+++ scripts/main.sh
@@ -1,8 +1,26 @@
 #!/bin/bash
 # Downloads a tarball artifact from AWS S3 and extracts it to the specified location.
 #
+# Usage: main.sh
+#
+# The following environment variables must be defined:
+#   - INPUT_NAME - The name of the artifact
+#   - INPUT_PATH - The path to be retrieved
+#   - INPUT_PATTERN - not implemented yet
+#   - INPUT_MERGE_MULTIPLE - not implemented yet
+#   - INPUT_REPOSITORY - the repository the artifact is associated with
+#   - INPUT_RUN_ID - the run ID the artifact is associated with
+#   - RUNNER_OS - the OS of the runne
+#   - ENV_S3_ARTIFACTS_BUCKET - the name of the AWS S3 bucket to use
+#   - ENV_AWS_ACCESS_KEY_ID - the AWS access key ID (optional if uploading to a public S3 bucket)
+#   - ENV_AWS_SECRET_ACCESS_KEY - the AWS secret access key (optional if uploading to a public S3 bucket)
+#   - DRY_RUN - whether to run without uploading to AWS (optional, set to true to enable dry run)
+#
 # based on open-turo/actions-s3-artifact
 # see: https://github.com/open-turo/actions-s3-artifact/blob/main/download/action.yaml
+
+# exit immediately if an error occurs
+set -e
 
 #region import scripts
 DIR="${BASH_SOURCE%/*}"
@@ -11,64 +29,77 @@ if [[ ! -d "$DIR" ]]; then DIR="$PWD"; fi
 #endregion
 
 #region read input arguments
-INPUT_NAME="$1"
-INPUT_PATH="$2"
-INPUT_PATTERN="$3"
-INPUT_MERGE_MULTIPLE="$4"
-INPUT_REPOSITORY="$5"
-INPUT_RUN_ID="$6"
-RUNNER_OS="$7"
-ENV_S3_ARTIFACTS_BUCKET="$8"
-ENV_AWS_ACCESS_KEY_ID="$9"
-ENV_AWS_SECRET_ACCESS_KEY="$10"
+echo "::debug::Inputs:"
+echo "::debug::    INPUT_NAME:                  $INPUT_NAME"
+echo "::debug::    INPUT_PATH:                  $INPUT_PATH"
+echo "::debug::    INPUT_PATTERN:               $INPUT_PATTERN"
+echo "::debug::    INPUT_MERGE_MULTIPLE:        $INPUT_MERGE_MULTIPLE"
+echo "::debug::    INPUT_REPOSITORY:            $INPUT_REPOSITORY"
+echo "::debug::    INPUT_RUN_ID:                $INPUT_RUN_ID"
+echo "::debug::    RUNNER_OS:                   $RUNNER_OS"
+echo "::debug::    ENV_S3_ARTIFACTS_BUCKET:     $ENV_S3_ARTIFACTS_BUCKET"
+echo "::debug::    ENV_AWS_ACCESS_KEY_ID:       $ENV_AWS_ACCESS_KEY_ID"
+echo "::debug::    ENV_AWS_SECRET_ACCESS_KEY:   $ENV_AWS_SECRET_ACCESS_KEY"
 #endregion
 
 #region validate input variables
 # validate script input variables
 if [[ "$INPUT_NAME" == "" ]]; then
     echo "::error::The values of 'NAME' input is not specified"
+    ERROR=true
 fi
 
 if [[ "$INPUT_PATH" == "" ]]; then
     echo "::error::The values of 'PATH' input is not specified"
-fi
-
-if [[ "$INPUT_PATTERN" == "" ]]; then
-    echo "::error::The values of 'INPUT_PATTERN' input is not specified"
+    ERROR=true
 fi
 
 if [[ "$INPUT_MERGE_MULTIPLE" == "" ]]; then
     echo "::error::The values of 'INPUT_MERGE_MULTIPLE' input is not specified"
+    ERROR=true
 fi
 
 if [[ "$INPUT_REPOSITORY" == "" ]]; then
     echo "::error::The values of 'INPUT_REPOSITORY' input is not specified"
+    ERROR=true
 fi
 
 if [[ "$INPUT_RUN_ID" == "" ]]; then
     echo "::error::The values of 'INPUT_RUN_ID' input is not specified"
+    ERROR=true
 fi
 
 # validate github actions variables
 if [[ "$RUNNER_OS" == "" ]]; then
     echo "::error::The values of 'RUNNER_OS' GitHub variable is not specified"
+    ERROR=true
 fi
 
-# check whether AWS credentials are specified and warn if they aren't
-if [[ "$ENV_AWS_ACCESS_KEY_ID" == "" || "$ENV_AWS_SECRET_ACCESS_KEY" == "" ]]; then
-    echo "::warn::AWS_ACCESS_KEY_ID and/or AWS_SECRET_ACCESS_KEY is missing from environment variables."
+if [[ "$DRY_RUN" != "true" ]]; then
+    # check whether AWS credentials are specified and warn if they aren't
+    if [[ "$ENV_AWS_ACCESS_KEY_ID" == "" || "$ENV_AWS_SECRET_ACCESS_KEY" == "" ]]; then
+        echo "::warn::AWS_ACCESS_KEY_ID and/or AWS_SECRET_ACCESS_KEY is missing from environment variables."
+        ERROR=true
+    fi
+
+    # check whether S3_ARTIFACTS_BUCKET is defined
+    if [[ "$ENV_S3_ARTIFACTS_BUCKET" == "" ]]; then
+        echo "::error::S3_ARTIFACTS_BUCKET is missing from environment variables."
+        ERROR=true
+    fi
 fi
 
-# check whether S3_ARTIFACTS_BUCKET is defined
-if [[ "$ENV_S3_ARTIFACTS_BUCKET" == "" ]]; then
-    echo "::error::S3_ARTIFACTS_BUCKET is missing from environment variables."
+if [[ "$ERROR" == "true" ]]; then
+    echo "::error::Input error(s) - exiting"
     exit 1
+else
+    echo "::debug::Validation complete"
 fi
 #endregion
 
 #region create temp directories
 # make sure that the path directory exists
-mkdir -p "$INPUTS_PATH"
+mkdir -p "$INPUT_PATH"
 
 # ensure we have a unique temporary directory to download to
 TMP_ARTIFACT="$RUNNER_TEMP/download-s3-artifact"
@@ -78,15 +109,17 @@ if [[ "$RUNNER_OS" == "Windows" ]]; then
     TMP_ARTIFACT=$(cygpath -u "$TMP_ARTIFACT")
 fi
 mkdir -p "$TMP_ARTIFACT"
+echo "::debug::The artifact directory is $TMP_ARTIFACT"
 
 # Create a unique directory for this particular action run
 TMPDIR="$(mktemp -d -p "$TMP_ARTIFACT" "download.XXXXXXXX")"
+mkdir -p "$TMPDIR"
 echo "::debug::Created temporary directory $TMPDIR"
 #endregion
 
 #region download artifact from AWS s3
 # Target for download, this can be a single file or a tarball
-TMPFILE="$TMPDIR/download-s3-artifact/artifact.tgz"
+TMPFILE="$TMPDIR/artifacts.tgz"
 
 # Get AWS S3 bucket URI and ensure it starts with "s3://"
 S3URI="$ENV_S3_ARTIFACTS_BUCKET"
@@ -102,7 +135,12 @@ S3URI="${S3URI%/}/$KEY"
 
 # Try to download
 echo "::debug::aws s3 cp '$S3URI' '$TMPFILE'"
-aws s3 cp "$S3URI" "$TMPFILE"
+if [[ "$DRY_RUN" == "true" ]]; then
+    # copy test file for testing
+    cp "./artifacts.tgz" "$TMPFILE"
+else
+    aws s3 cp "$S3URI" "$TMPFILE"
+fi
 echo "::debug::File downloaded successfully to $TMPFILE"
 #endregion
 
@@ -113,23 +151,31 @@ echo "::debug::File downloaded successfully to $TMPFILE"
 # fi
 
 # Downloaded a tarball, extract it
-# TODO: Should we check the path input to make sure it is in the project?
-echo "::debug::tar -xzvf '$TMPFILE' -C '$INPUTS_PATH' $TAR_CLI_ARGS"
-tar -xzvf "$TMPFILE" -C "$INPUTS_PATH" $TAR_CLI_ARGS
+# TODO: Should we check the path input to make sure it exists?
+echo "::debug::tar -xzvf '$TMPFILE' -C '$INPUT_PATH' $TAR_CLI_ARGS"
+tar -xzvf "$TMPFILE" -C "$INPUT_PATH" $TAR_CLI_ARGS
 
+# list out everything in the extracted location
 if [[ -n "$RUNNER_DEBUG" ]]; then
-    echo "::debug::Contents of artifact path"
-    echo "$(tree -a '$INPUTS_PATH' 2>&1)"
+    echo "::debug::Contents of our temporary directory"
+    if [[ "$RUNNER_OS" = "Windows" ]]; then
+        # TODO: Can I make this debug somehow?
+        cmd //c tree //f "$INPUT_PATH"
+    else
+        echo "::debug::$(tree -a "$INPUT_PATH" 2>&1)"
+    fi
 fi
 #endregion
 
 #region generate outputs
 # set output
 # TODO: I don't think this output is correct. Need to investigate.
-echo "download-path='$INPUTS_PATH'" >>$GITHUB_OUTPUT
+echo "download-path='$INPUT_PATH'" >> $GITHUB_OUTPUT
 #endregion
 
 #region clean up temp dir
-# clean up temp files
-rm -rf $TMP_ARTIFACT
+# TODO: move to clean up step?
+if [[ "$DRY_RUN" != "true" ]]; then
+    rm -rf $TMP_ARTIFACT
+fi
 #endregion
